Fix: Nomenclature confusion
