+++ packages/toolpad-app/src/toolpad/AppEditor/NoPageFound.tsx
@@ -1,7 +1,7 @@
 import { Grid, Typography, Button } from '@mui/material';
 import * as React from 'react';
 import useBoolean from '../../utils/useBoolean';
-import CreatePageNodeDialog from './HierarchyExplorer/CreatePageNodeDialog';
+import CreatePageNodeDialog from './PagesExplorer/CreatePageNodeDialog';
 
 export default function NoPageFound() {
   const {
