+++ packages/toolpad-app/src/toolpad/AppEditor/NodeNameEditor.tsx
@@ -2,7 +2,7 @@ import { SxProps, TextField } from '@mui/material';
 import * as React from 'react';
 import * as appDom from '../../appDom';
 import { useDom, useDomApi } from '../AppState';
-import { useNodeNameValidation } from './HierarchyExplorer/validation';
+import { useNodeNameValidation } from './PagesExplorer/validation';
 import client from '../../api';
 
 interface NodeNameEditorProps {
